Upgrade to gulp 4, babel 7 and eslint 6
