+++ .babelrc
@@ -1,5 +1,5 @@
 {
   "presets": [
-    "latest"
+    "@babel/preset-env"
   ]
 }
